Domestic_Violence_Data
This CSV file contains the socio-demographic and economic data used for the Domestic Violence Analysis project. The dataset includes features such as education level, marital status, age, income, and other relevant factors that help in understanding the incidence and patterns of domestic violence. The data is utilized to analyze correlations and train models to predict the likelihood of domestic violence based on these socio-demographic attributes.